Refactored to remove else statements
Hey everyone, just a random stranger on the interwebs. 

I popped into this class to practice refactoring and I noticed a few else statements that could be removed. If you like the refactor please consider including it, if not then just let me know and I'll close the PR. Thanks =^D